SemanticKernel/requirements.txt
semantic-kernel[azure]==1.34.0
openai==1.97.0
python-dotenv==1.0.1
